+++ core/src/com/ray3k/skincomposer/MainListener.java
@@ -57,6 +57,7 @@ import com.ray3k.skincomposer.data.DrawableData;
 import com.ray3k.skincomposer.data.FreeTypeFontData;
 import com.ray3k.skincomposer.dialog.DialogCustomClass.CustomClassListener;
 import com.ray3k.skincomposer.dialog.DialogCustomStyle;
+import com.ray3k.skincomposer.dialog.DialogListener;
 import com.ray3k.skincomposer.dialog.DialogWelcome.WelcomeListener;
 import com.ray3k.skincomposer.utils.Utils;
 import java.io.File;
@@ -70,6 +71,7 @@ public class MainListener extends RootTableListener {
     private final JsonData jsonData;
     private final Main main;
     private WelcomeDialogListener welcomeListener;
+    private DialogListener dialogListener;
     
     public MainListener(Main main) {
         this.root = main.getRootTable();
@@ -78,12 +80,24 @@ public class MainListener extends RootTableListener {
         this.projectData = main.getProjectData();
         this.jsonData = main.getProjectData().getJsonData();
         this.main = main;
+        
+        dialogListener = new DialogListener() {
+            @Override
+            public void opened() {
+                main.getDesktopWorker().removeFilesDroppedListener(main.getRootTable().getFilesDroppedListener());
+            }
+
+            @Override
+            public void closed() {
+                main.getDesktopWorker().addFilesDroppedListener(main.getRootTable().getFilesDroppedListener());
+            }
+        };
     }
     
     public void createWelcomeListener() {
         welcomeListener = new WelcomeDialogListener();
         if (main.getProjectData().isAllowingWelcome()) {
-            dialogFactory.showWelcomeDialog(welcomeListener);
+            dialogFactory.showWelcomeDialog(welcomeListener, dialogListener);
         }
     }
     
@@ -97,7 +111,7 @@ public class MainListener extends RootTableListener {
                 openFile();
                 break;
             case RECENT_FILES:
-                dialogFactory.recentFiles();
+                dialogFactory.recentFiles(dialogListener);
                 break;
             case SAVE:
                 saveFile(null);
@@ -106,7 +120,7 @@ public class MainListener extends RootTableListener {
                 saveAsFile(null);
                 break;
             case WELCOME:
-                main.getDialogFactory().showWelcomeDialog(getWelcomeListener());
+                main.getDialogFactory().showWelcomeDialog(getWelcomeListener(), dialogListener);
                 break;
             case IMPORT:
                 importFile();
@@ -115,7 +129,7 @@ public class MainListener extends RootTableListener {
                 exportFile();
                 break;
             case EXIT:
-                dialogFactory.showCloseDialog();
+                dialogFactory.showCloseDialog(dialogListener);
                 break;
             case UNDO:
                 main.getUndoableManager().undo();
@@ -124,19 +138,19 @@ public class MainListener extends RootTableListener {
                 main.getUndoableManager().redo();
                 break;
             case SETTINGS:
-                dialogFactory.showSettings();
+                dialogFactory.showSettings(dialogListener);
                 break;
             case COLORS:
-                dialogFactory.showDialogColors();
+                dialogFactory.showDialogColors(dialogListener);
                 break;
             case FONTS:
-                dialogFactory.showFonts();
+                dialogFactory.showFonts(dialogListener);
                 break;
             case DRAWABLES:
-                dialogFactory.showDialogDrawables();
+                dialogFactory.showDialogDrawables(dialogListener);
                 break;
             case ABOUT:
-                dialogFactory.showAbout();
+                dialogFactory.showAbout(dialogListener);
                 break;
             case CLASS_SELECTED:
                 updateStyleProperties();
@@ -270,7 +284,7 @@ public class MainListener extends RootTableListener {
     
     public void newFile() {
         if (!projectData.areChangesSaved() && !projectData.isNewProject()) {
-            dialogFactory.yesNoCancelDialog("Save Changes?",
+            var dialog = dialogFactory.yesNoCancelDialog("Save Changes?",
                     "Do you want to save changes to the existing project?"
                             + "\nAll unsaved changes will be lost.",
                     (int selection) -> {
@@ -281,7 +295,18 @@ public class MainListener extends RootTableListener {
                         } else if (selection == 1) {
                             projectData.clear();
                         }
-                    });
+                    }, dialogListener);
+            dialog.addListener(new DialogListener() {
+                @Override
+                public void opened() {
+                    main.getDesktopWorker().removeFilesDroppedListener(main.getRootTable().getFilesDroppedListener());
+                }
+
+                @Override
+                public void closed() {
+                    main.getDesktopWorker().addFilesDroppedListener(main.getRootTable().getFilesDroppedListener());
+                }
+            });
         } else {
             projectData.clear();
         }
@@ -312,7 +337,7 @@ public class MainListener extends RootTableListener {
         };
         
         if (!projectData.areChangesSaved() && !projectData.isNewProject()) {
-            dialogFactory.yesNoCancelDialog("Save Changes?",
+            var dialog = dialogFactory.yesNoCancelDialog("Save Changes?",
                     "Do you want to save changes to the existing project?"
                     + "\nAll unsaved changes will be lost.",
                     (int selection) -> {
@@ -321,7 +346,19 @@ public class MainListener extends RootTableListener {
                         } else if (selection == 1) {
                             dialogFactory.showDialogLoading(runnable);
                         }
-                    });
+                    }, dialogListener);
+            
+            dialog.addListener(new DialogListener() {
+                @Override
+                public void opened() {
+                    main.getDesktopWorker().removeFilesDroppedListener(main.getRootTable().getFilesDroppedListener());
+                }
+
+                @Override
+                public void closed() {
+                    main.getDesktopWorker().addFilesDroppedListener(main.getRootTable().getFilesDroppedListener());
+                }
+            });
         } else {
             dialogFactory.showDialogLoading(runnable);
         }
@@ -350,7 +387,7 @@ public class MainListener extends RootTableListener {
         };
         
         if (!projectData.areChangesSaved() && !projectData.isNewProject()) {
-            dialogFactory.yesNoCancelDialog("Save Changes?",
+            var dialog = dialogFactory.yesNoCancelDialog("Save Changes?",
                     "Do you want to save changes to the existing project?"
                     + "\nAll unsaved changes will be lost.",
                     (int selection) -> {
@@ -359,7 +396,19 @@ public class MainListener extends RootTableListener {
                         } else if (selection == 1) {
                             dialogFactory.showDialogLoading(runnable);
                         }
-                    });
+                    }, dialogListener);
+            
+            dialog.addListener(new DialogListener() {
+                @Override
+                public void opened() {
+                    main.getDesktopWorker().removeFilesDroppedListener(main.getRootTable().getFilesDroppedListener());
+                }
+
+                @Override
+                public void closed() {
+                    main.getDesktopWorker().addFilesDroppedListener(main.getRootTable().getFilesDroppedListener());
+                }
+            });
         } else {
             dialogFactory.showDialogLoading(runnable);
         }
@@ -506,11 +555,11 @@ public class MainListener extends RootTableListener {
     public void stylePropertyChanged(StyleProperty styleProperty,
             Actor styleActor) {
         if (styleProperty.type == Drawable.class) {
-            dialogFactory.showDialogDrawables(styleProperty);
+            dialogFactory.showDialogDrawables(styleProperty, dialogListener);
         } else if (styleProperty.type == Color.class) {
-            dialogFactory.showDialogColors(styleProperty);
+            dialogFactory.showDialogColors(styleProperty, dialogListener);
         } else if (styleProperty.type == BitmapFont.class) {
-            dialogFactory.showDialogFonts(styleProperty);
+            dialogFactory.showDialogFonts(styleProperty, dialogListener);
         } else if (styleProperty.type == Float.TYPE) {
             main.getUndoableManager().addUndoable(new UndoableManager.DoubleUndoable(main, styleProperty, ((Spinner) styleActor).getValue()), false);
         } else if (styleProperty.type == ScrollPaneStyle.class) {
@@ -621,13 +670,13 @@ public class MainListener extends RootTableListener {
             Actor styleActor) {
         if (null != customProperty.getType()) switch (customProperty.getType()) {
             case DRAWABLE:
-                dialogFactory.showDialogDrawables(customProperty);
+                dialogFactory.showDialogDrawables(customProperty, dialogListener);
                 break;
             case COLOR:
-                dialogFactory.showDialogColors(customProperty);
+                dialogFactory.showDialogColors(customProperty, dialogListener);
                 break;
             case FONT:
-                dialogFactory.showDialogFonts(customProperty);
+                dialogFactory.showDialogFonts(customProperty, dialogListener);
                 break;
             case NUMBER:
                 main.getUndoableManager().addUndoable(new UndoableManager.CustomDoubleUndoable(main, customProperty, ((Spinner) styleActor).getValue()), false);
@@ -642,6 +691,11 @@ public class MainListener extends RootTableListener {
             default:
                 break;
         }  
+    }
+
+    @Override
+    public void droppedScmpFile(FileHandle fileHandle) {
+        openFile(fileHandle);
     }
     
     public void refreshTextureAtlas() {